Implement dynamic profile routing with friend actions and profile page
